+++ src/main/webapp/resources/app/components/place/place.template.html
@@ -95,14 +95,14 @@
 					</div>
 				</div>
 
-				<div class="col s12 m5 l5 grid-item card white"
+				<div class="col s12 m4 l5 grid-item card white"
 					ng-repeat="(i, feedback) in feedbacks" less="true" ng-href="#{{i}}!">
 					<div class="card-content grey-text text-darken-1">
 					<img class="circle" ng-src="{{feedback.user.userpic}}" />
 						<span class="card-title"> <span>{{feedback.user.username}}</span></br>
 							<span class="rel_date" ng-bind="date | date:'MM/dd/yyyy'">{{feedback.date}}</span>
 						</span><br />
-						<p class="less">{{feedback.body.substr(0, 400)}}</p>
+						<p class="less">{{feedback.body.substr(0, 440)}}</p>
 						<p class="more hidden">{{feedback.body}}</p>
 					</div>
 					<div class="card-action">
@@ -112,6 +112,79 @@
 				</div>
 			</div>
 			<div class="place-overlay">&nbsp;</div>
+			
+			  <!-- Modal Structure for Feedback About Place: -->
+		<div id="feedback-modal" class="modal modal-fixed-footer mymodal"
+			ng-controller="PlaceFeedbackController">
+			<div class="modal-content">
+				<span class="card-title feedback-username"> <img
+					class="circle feedback-userpic" ng-src="{{userpicture.userpic}}" /> <span>{{username.username}}</span>
+				</span><br />
+				<div class="input-field">
+					<div class="typeit">Type your feedback</div>
+					<textarea placeholder=" " id="typing-feedback" type="text"
+						class="validate" ng-model="text"></textarea>
+					<label for="typing-feedback" class="typeit"></label>
+				</div>
+				<!-- <span ng-bind="text"></span> -->
+				<form action="#">
+					<div class="file-field input-field images-uploader">
+						<div class="commenting-button">
+							<span><i
+								class="material-icons camera-icon grey-text text-lighten-1">photo_camera</i></span>
+							<input type="file">
+						</div>
+						<div class="file-path-wrapper">
+							<input class="file-path validate" type="text">
+						</div>
+					</div>
+				</form>
+			</div>
+			<div class="modal-footer">
+				<button type="button"
+					class="bttns modal-action modal-close waves-effect waves-green btn-flat"
+					ng-click="addText()">Publish</button> <button type="button" 
+					class="bttns modal-action modal-close waves-effect waves-red btn-flat">Cancel</button>
+			</div>
+		</div>
+		
+		<!-- Modal Structures of Edit Place: -->
+		<div id="name-modal" class="modal modal-fixed-footer mymodal">
+			<div class="modal-content">
+				<span class="card-title feedback-username"> <span></span>
+				</span><br />
+				<div class="input-field">
+					<label for="" class="place_name">Place Name:</label>
+					<input type="text" placeholder="" id=""
+						class="validate" ng-model="name" />
+				</div>
+			</div>
+			<div class="modal-footer">
+				<button type="button"
+					class="bttns modal-action modal-close waves-effect waves-green btn-flat"
+					ng-click="update_name()">Save</button> <button type="button" 
+					class="bttns modal-action modal-close waves-effect waves-red btn-flat">Cancel</button>
+			</div>
+		</div>
+
+		<div id="description-modal" class="modal modal-fixed-footer mymodal">
+			<div class="modal-content">
+				<span class="card-title feedback-username"> <span></span>
+				</span><br />
+				<div class="input-field">
+				    <label for="" class="place_name">Place Description:</label>
+					<input type="text" placeholder="" id=""
+						class="validate" ng-model="description" />
+				</div>
+			</div>
+			<div class="modal-footer">
+				<button type="button"
+					class="bttns modal-action modal-close waves-effect waves-green btn-flat"
+					ng-click="update_description()">Save</button> <button type="button" 
+					class="bttns modal-action modal-close waves-effect waves-red btn-flat">Cancel</button>
+			</div>
+		</div>
+		
 		</div>
 
 	</div>
